Remove animejd
