+++ libs/island-ui/core/src/lib/LoadingIcon/LoadingIcon.tsx
@@ -1,7 +1,6 @@
-import React, { forwardRef, useEffect, useRef } from 'react'
+import React, { forwardRef } from 'react'
 import { Colors, theme } from '@island.is/island-ui/theme'
-import anime, { AnimeInstance } from 'animejs'
-import { motion, useAnimation } from 'framer-motion'
+import AnimatedLine from './AnimatedLine'
 
 interface LoadingIconProps {
   animate?: boolean
@@ -11,80 +10,6 @@ interface LoadingIconProps {
 
 export const LoadingIcon = forwardRef<SVGSVGElement, LoadingIconProps>(
   ({ animate = true, color, size }, ref) => {
-    const animationRef = useRef<AnimeInstance | null>(null)
-    const lineRef = useRef<SVGLineElement | null>(null)
-    const controls = useAnimation()
-    useEffect(() => {
-      seq()
-    }, [])
-    const seq = async () => {
-      await controls.start({
-        x1: 50,
-        transition: { duration: 0.15, ease: 'easeInOutSine' },
-      })
-      await controls.start({
-        x2: 50,
-        transition: { duration: 0.15, ease: 'easeInOutSine' },
-      })
-      await controls.start({
-        x1: 88,
-        transition: { duration: 0.15, ease: 'easeInOutSine' },
-      })
-      await controls.start({
-        x2: 88,
-        transition: { duration: 0.15, ease: 'easeInOutSine' },
-      })
-      await controls.start({
-        y1: 50,
-        transition: { duration: 0.15, ease: 'easeInOutSine' },
-      })
-      await controls.start({
-        y2: 50,
-        transition: { duration: 0.15, ease: 'easeInOutSine' },
-      })
-      await controls.start({
-        y1: 88,
-        transition: { duration: 0.15, ease: 'easeInOutSine' },
-      })
-      await controls.start({
-        y2: 88,
-        transition: { duration: 0.15, ease: 'easeInOutSine' },
-      })
-      await controls.start({
-        x1: 50,
-        transition: { duration: 0.15, ease: 'easeInOutSine' },
-      })
-      await controls.start({
-        x2: 50,
-        transition: { duration: 0.15, ease: 'easeInOutSine' },
-      })
-      await controls.start({
-        x1: 12,
-        transition: { duration: 0.15, ease: 'easeInOutSine' },
-      })
-      await controls.start({
-        x2: 12,
-        transition: { duration: 0.15, ease: 'easeInOutSine' },
-      })
-      await controls.start({
-        y1: 50,
-        transition: { duration: 0.15, ease: 'easeInOutSine' },
-      })
-      await controls.start({
-        y2: 50,
-        transition: { duration: 0.15, ease: 'easeInOutSine' },
-      })
-      await controls.start({
-        y1: 12,
-        transition: { duration: 0.15, ease: 'easeInOutSine' },
-      })
-      await controls.start({
-        y2: 12,
-        transition: { duration: 0.15, ease: 'easeInOutSine' },
-      })
-      seq()
-    }
-
     const usedColor = color
       ? theme.color[color]
       : `url(#loading-icon-new-linear-gradient)`
@@ -120,19 +45,7 @@ export const LoadingIcon = forwardRef<SVGSVGElement, LoadingIconProps>(
         <circle cx="12" cy="88" r="12" fill={usedColor} />
         <circle cx="50" cy="88" r="12" fill={usedColor} />
         <circle cx="88" cy="88" r="12" fill={usedColor} />
-        {!!animate && (
-          <motion.line
-            cx={500}
-            animate={controls}
-            x1="12"
-            x2="12"
-            y1="12"
-            y2="12"
-            stroke={usedColor}
-            strokeWidth="24"
-            strokeLinecap="round"
-          />
-        )}
+        {!!animate && <AnimatedLine color={color} />}
       </svg>
     )
   },
