+++ package.json
@@ -90,7 +90,6 @@
     "@types/stream-buffers": "3.0.3",
     "@xstate/react": "0.8.1",
     "amazon-s3-uri": "0.0.3",
-    "animejs": "3.2.1",
     "apollo-cache-inmemory": "1.6.6",
     "apollo-client": "2.6.8",
     "apollo-datasource-rest": "0.9.2",
@@ -304,7 +303,6 @@
     "@testing-library/react": "11.1.2",
     "@testing-library/react-hooks": "5.0.3",
     "@testing-library/user-event": "12.1.5",
-    "@types/animejs": "3.1.2",
     "@types/aws-sdk": "2.7.0",
     "@types/aws4": "1.5.1",
     "@types/bull": "3.14.2",
